+++ token/services/vault/rws/translator/translator_test.go
@@ -10,11 +10,9 @@ import (
 	"strconv"
 
 	"github.com/hyperledger-labs/fabric-token-sdk/token/services/ttx"
-
+	"github.com/hyperledger-labs/fabric-token-sdk/token/services/vault/rws/keys"
 	"github.com/hyperledger-labs/fabric-token-sdk/token/services/vault/rws/translator"
 	mock2 "github.com/hyperledger-labs/fabric-token-sdk/token/services/vault/rws/translator/mock"
-
-	"github.com/hyperledger-labs/fabric-token-sdk/token/services/vault/rws/keys"
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 	"github.com/pkg/errors"
